+++ src/player.js
@@ -1,27 +1,27 @@
 class Player extends Entity {
-    constructor(scene, x = 250, y = 550, animation = 'player', health = 100, gunType='player') {
-        super(scene, x, y, animation, health, gunType);
-        this.scene.physics.add.collider(this.sprite, enemyBullets.sprites, this.hit);
-        this.scene.physics.add.collider(this.sprite, entities.sprites, this.hit);
+    constructor(x = 250, y = 550, animation = 'player', health = 100, gunType='player') {
+        super(x, y, animation, health, gunType);
+        scene.physics.add.collider(this.sprite, world.enemyBullets.sprites, this.hit);
+        scene.physics.add.collider(this.sprite, world.entities.sprites, this.hit);
         this.sprite.body.collideWorldBounds = true;
         this.currentGunType = gunType;
         this.changeGunType = gunType;
-        this.weapon = new (stringToWeapon(gunType))(playerBullets, 1, this);
+        this.weapon = new (stringToWeapon(gunType))(world.playerBullets, 1, this);
         this.sprite.name = 'player';
     }
     move() {
         if (this.currentGunType !== this.changeGunType){
-            this.weapon = new (stringToWeapon(this.changeGunType))(playerBullets, 1, this);
+            this.weapon = new (stringToWeapon(this.changeGunType))(world.playerBullets, 1, this);
             this.currentGunType = this.changeGunType;
         }
-        if (this.scene.input.keyboard.keys[Phaser.Input.Keyboard.KeyCodes.A].isDown) {
+        if (scene.input.keyboard.keys[Phaser.Input.Keyboard.KeyCodes.A].isDown) {
             if(this.sprite.body.velocity.x > 0){
                 this.sprite.setAccelerationX(0);
                 this.sprite.setVelocityX(0);
             } else {
                 this.sprite.setAccelerationX(-500);
             }
-        } else if (this.scene.input.keyboard.keys[Phaser.Input.Keyboard.KeyCodes.D].isDown) {
+        } else if (scene.input.keyboard.keys[Phaser.Input.Keyboard.KeyCodes.D].isDown) {
             if(this.sprite.body.velocity.x < 0){
                 this.sprite.setAccelerationX(0);
                 this.sprite.setVelocityX(0);
@@ -36,8 +36,8 @@ class Player extends Entity {
                 this.sprite.setAccelerationX(-1*Math.sign(this.sprite.body.velocity.x)*500);
             }
         }
-        if (!win){
-            if (this.scene.input.keyboard.keys[Phaser.Input.Keyboard.KeyCodes.SPACE].isDown) {
+        if (!world.win){
+            if (scene.input.keyboard.keys[Phaser.Input.Keyboard.KeyCodes.SPACE].isDown) {
                 this.weapon.shoot();
             }
         }
@@ -45,10 +45,10 @@ class Player extends Entity {
     winMove(){
         if (this.sprite.y < -50) {
             this.sprite.destroy();
-            entities.removeElement(this);
-            playerHealthBar.destroy();
-            playerReloadBar.destroy();
-            scoreText.destroy();
+            world.entities.removeElement(this);
+            world.playerHealthBar.destroy();
+            world.playerReloadBar.destroy();
+            world.scoreText.destroy();
         } else {
             if (Math.abs(this.sprite.x-250)>10){
                 this.sprite.setVelocityX(Math.sign(250-this.sprite.x)*100);